# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CareConnect API Documentation",
        "description": "API documentation for CareConnect Healthcare Management System",
        "version": "1.0.0",
        "contact": {
            "email": "blissbonheur1@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "API Server V1"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object"
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Success message"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": ["patient", "doctor", "admin"]
                    }
                }
            }
        }
    },
    "paths": {
...
        "/api/user/password": {
            "put": {
                "tags": ["Users"],
                "summary": "Update Authenticated User Password",
                "security": [{"bearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["current_password", "new_password"],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "description": "Current password"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "description": "New password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/delete/{id}": {
            "delete": {
                "tags": ["Users"],
                "summary": "Delete a User Account",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "User ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User account deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/user/signup": {
            "post": {
                "tags": ["Users"],
                "summary": "Create a User Account",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["username", "firstName", "lastName", "email", "password", "role", "gender", "dateOfBirth"],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "description": "User's username"
                                    },
                                    "firstName": {
                                        "type": "string",
                                        "description": "User's first name"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "description": "User's last name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "description": "User's email address"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "description": "User's password"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": ["patient", "doctor", "admin"],
                                        "description": "User's role"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": ["Male", "Female", "Other"],
                                        "description": "User's gender"
                                    },
                                    "dateOfBirth": {
                                        "type": "string",
                                        "format": "date",
                                        "description": "User's date of birth (YYYY-MM-DD)"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "description": "User's phone number"
                                    },
                                    "address": {
                                        "type": "string",
                                        "description": "User's address"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/all": {
            "get": {
                "tags": ["Users"],
                "summary": "List all User Accounts",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all user accounts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/profile/{id}": {
            "put": {
                "tags": ["Users"],
                "summary": "Update User Account",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "User ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "description": "Updated first name"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "description": "Updated last name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "description": "Updated email"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "description": "Updated phone number"
                                    },
                                    "address": {
                                        "type": "string",
                                        "description": "Updated address"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User account updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}